[UPD] Use connector unit
Readapt the code with connector unit
