+++ cmis_write/ir_attachment.py
@@ -22,8 +22,7 @@
 
 from openerp.osv import orm, osv, fields
 from openerp.addons.connector.session import ConnectorSession
-from openerp.addons.connector.queue.job import job
-import base64
+from .unit.import_synchronizer import create_doc_in_edm
 from openerp import SUPERUSER_ID
 from openerp.tools.translate import _
 from openerp.tools import ustr
@@ -130,8 +129,8 @@ class ir_attachment(orm.Model):
                                 ('storing_ok', '=', 'True')
                             ], context=context)
                         # Connect to the backend
-                        repo = cmis_backend_obj._auth(
-                            cr, uid, backend, context=context)
+                        repo = cmis_backend_obj.check_auth(
+                            cr, uid, [backend], context=context)
                         # Get the id of document
                         id_dms = doc.object_doc_id
                         # Get results from id of document
@@ -229,66 +228,3 @@ class ir_attachment(orm.Model):
             help="Records"
         ),
     }
-
-
-@job
-def create_doc_in_edm(session, model_name, value, res,
-                      dict_metadata, user_login, filters=None):
-    """
-    This method allows to create a doc from Odoo to DMS
-    """
-    cr = session.cr
-    uid = session.uid
-    context = session.context
-    if context is None:
-        context = {}
-    ir_attach_obj = session.pool.get('ir.attachment')
-    ir_attach_doc_backend_obj = session.pool.get('ir.attachment.doc.backend')
-    cmis_backend_obj = session.pool.get('cmis.backend')
-
-    # List of backend with storing_ok is True
-    ids = cmis_backend_obj.search(
-        cr, uid, [('storing_ok', '=', 'True')], context=context)
-
-    for backend in session.browse('cmis.backend', ids):
-        repo = cmis_backend_obj._auth(cr, uid, backend.id, context=context)
-        root = repo.rootFolder
-        folder_path = backend.initial_directory_write
-        # Document properties
-        if value.get('name'):
-            file_name = value.get('name')
-        elif value.get('datas_fname'):
-            file_name = value.get('datas_fname')
-        else:
-            file_name = value.get('datas_fname')
-        props = {
-            'cmis:name': file_name,
-            'cmis:description': value.get('description'),
-            'cmis:createdBy': user_login,
-        }
-        # Add list of metadata in props
-        if len(dict_metadata):
-            for k, v in dict_metadata.iteritems():
-                props[k] = v
-        if folder_path:
-            sub1 = repo.getObjectByPath(folder_path)
-        else:
-            sub1 = root
-        someDoc = sub1.createDocumentFromString(
-            file_name,
-            contentString=base64.b64decode(
-                value.get('datas')), contentType=value.get('file_type')
-        )
-        # TODO: create custom properties on a document (Alfresco)
-        # someDoc.getProperties().update(props)
-        # Updating ir.attachment object with the new id
-        # of document generated by DMS
-        ir_attach_obj.write(cr, uid, res, {
-            'id_dms': someDoc.getObjectId()}, context=context)
-        ir_attach_doc_backend_obj.create(
-            cr, uid, {
-                'attachment_id': res,
-                'backend_id': backend.id,
-                'object_doc_id': someDoc.getObjectId(),
-            }, context=context)
-    return True
